+++ packages/components/molecules/f-card-with-content/CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v3.3.0
+------------------------------
+*August 3, 2022*
+
+### Added
+- Node 16 compatible version of `@justeat/f-vue-icons`.
+
+
 v3.2.0
  ------------------------------
  *July 19, 2022*