+++ packages/components/organisms/f-cookie-banner/CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.5.0
+------------------------------
+*August 04, 2022*
+
+### Fixed
+- Node 16 issue around window.dispatchEvent.
+
+
+v4.4.0
+------------------------------
+*July 29, 2022*
+
+### Added
+- Node 16 compatible version of `@justeat/f-vue-icons`.
+
+
 v4.3.0
  ------------------------------
  *July 26, 2022*