+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -2,6 +2,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v10.5.0
+------------------------------
+*July 29, 2022*
+
+### Added
+- Node 16 compatible version of `@justeat/f-popover`.
+- Node 16 compatible version of `@justeat/f-button`.
+
+
 v10.4.0
 ------------------------------
 *July 26, 2022*