+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.55.0
+------------------------------
+*August 10, 2022*
+
+### Changed
+- `storybook:deploy` step to work in GitHub Actions.
+
+
 v0.54.1
 ------------------------------
 *July 26, 2022*